complete basic distribution
